requirements.txt
asgiref==3.4.1
Django==3.2.7
django-summernote==0.8.11.6
django-tinymce==3.3.0
pkg_resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
